Cargo fmt, rename fonts to font_cache
